refactor model input types
